requirements.txt
dbt-snowflake==1.7.0
snowflake-connector-python
python-dotenv